fixed splash page not rendering issue by passing a prop manually from app.js
components are set to only return values if they are selected by useState.  Usestates starting value is set to homebutton but because useState is asynchronous its initial value was only defined after the rest of the javascript executed.  Passed in a custom prop setting a hard value for userpageselection if the usestate value had not already selected a valid option